chore(async-rewriter2): split into one file per stage
Michael suggested doing this, and I agreed but forgot to actually do it.
