+++ packages/async-rewriter2/src/async-writer-babel.ts
@@ -1,12 +1,13 @@
-/* eslint no-sync: 0, no-console:0, complexity:0, dot-notation: 0 */
+/* eslint-disable no-sync */
 import * as babel from '@babel/core';
-import * as BabelTypes from '@babel/types';
 import runtimeSupport from './runtime-support.nocov';
+import wrapAsFunctionPlugin from './stages/wrap-as-iife';
+import makeMaybeAsyncFunctionPlugin from './stages/transform-maybe-await';
 
 /**
- * General notes for this file:
+ * General notes for this package:
  *
- * This file contains two babel plugins used in async rewriting, plus a helper
+ * This package contains two babel plugins used in async rewriting, plus a helper
  * to apply these plugins to plain code.
  *
  * If you have not worked with babel plugins,
@@ -15,689 +16,12 @@ import runtimeSupport from './runtime-support.nocov';
  *
  * https://astexplorer.net/ is also massively helpful, and enables working
  * with babel plugins as live transformers. It doesn't understand all TS syntax,
- * but pasting the compiled output from lib/async-writer-babel.js is already
+ * but pasting the compiled output from lib/stages/[...].js is already
  * a very good start.
  *
  * The README file for this repository contains a more high-level technical
  * overview.
  */
-
-/**
- * First-step transform plugin: Wrap the entire program inside a function,
- * and re-export variables.
- *
- * ```js
- * function foo() { return db.test.find(); }
- * class A {}
-* foo()```
- *
- * is converted into roughly:
- * ```js
- * var A;
- *
- * function foo() {
- *   return db.test.find();
- * }
- *
- * (() => {
- *   A = class A {};
- *   return foo();
- * })();
- * ```
- *
- * This is necessary because the second transformation step only works with
- * functions, not top-level program code.
- *
- * It is also 'nice' in the sense that it converts 'let' and 'const' variable
- * declarations into 'var' declarations, which is usually not something one
- * wants to do in JS, but for REPLs it actually increases usability a lot.
- */
-type WrapState = {
-  // All statements moved from the Program-level scope into the new function.
-  movedStatements: babel.types.Statement[];
-  // All function declarations. These have to be moved to the outermost scope
-  // in order to allow for proper hoisting semantics.
-  functionDeclarations: babel.types.FunctionDeclaration[];
-  // Whether the function-creating part has completed.
-  hasFinishedMoving: boolean;
-  // A list of all variables names that have been moved to outside of the function.
-  variables: string[];
-  // Whether the conversion of completion records into return statements already took place.
-  addedCompletionRecords: boolean;
-  // The identifier used for returning the completion record
-  completionRecordId: babel.types.Identifier;
-};
-export const wrapAsFunctionPlugin = ({ types: t }: { types: typeof BabelTypes }): babel.PluginObj<WrapState> => {
-  function asNodeKey(v: any): keyof babel.types.Node { return v; }
-  const excludeFromCompletion = asNodeKey(Symbol('excludedFromCompletion'));
-  return {
-    pre() {
-      this.movedStatements = [];
-      this.functionDeclarations = [];
-      this.hasFinishedMoving = false;
-      this.addedCompletionRecords = false;
-      this.variables = [];
-    },
-    visitor: {
-      Statement(path): void {
-        if (this.hasFinishedMoving) return;
-        if (path.isDeclaration() && !path.getFunctionParent()) {
-          // The complicated case: We've encountered a variable/function/class
-          // declaration.
-          if (path.isVariableDeclaration()) {
-            if (path.parentPath.isProgram() || path.node.kind === 'var') {
-              // We turn variables into outermost-scope variables if they are
-              // a) hoisted 'var's with no function parent or b) let/const
-              // at the top-level (i.e. no block scoping).
-              const asAssignments = [];
-              for (const decl of path.node.declarations) {
-                // Copy variable names.
-                this.variables.push((decl.id as babel.types.Identifier).name);
-                if (decl.init !== null) {
-                  // If there is an initializer for this variable, turn it into
-                  // an assignment expression.
-                  const expr = t.assignmentExpression('=', decl.id, decl.init);
-                  if (path.node.kind !== 'var') {
-                    Object.assign(expr, { [excludeFromCompletion]: true });
-                  }
-                  asAssignments.push(t.expressionStatement(expr));
-                }
-              }
-              if (path.parentPath.isProgram()) {
-                this.movedStatements.push(...asAssignments);
-                path.remove();
-              } else {
-                path.replaceWithMultiple(asAssignments);
-              }
-              return;
-            }
-          } else if (path.isFunctionDeclaration()) {
-            // Move top-level functions separately for hoisting.
-            this.functionDeclarations.push(path.node);
-            if (path.node.id) {
-              path.replaceWith(t.expressionStatement(path.node.id));
-            } else {
-              // Unsure how to reach this, but babel says the type of
-              // FunctionDeclaration['id'] is 'Identifier | null'.
-              path.remove();
-            }
-            return;
-          } else if (path.isClassDeclaration() && path.parentPath.isProgram()) {
-            // Convert this declaration into an assignment expression, i.e.
-            // `class A {}` becomes `A = class A {}`.
-            // Unlike `var`
-            this.variables.push(path.node.id.name);
-            this.movedStatements.push(
-              t.expressionStatement(
-                t.assignmentExpression('=', path.node.id,
-                  t.classExpression(
-                    path.node.id, path.node.superClass, path.node.body))));
-            path.replaceWith(t.expressionStatement(path.node.id));
-            return;
-          }
-        }
-        // All other declarations are currently either about TypeScript
-        // or ES modules. We treat them like non-declaration statements here
-        // and move them into the generated IIFE.
-        if (path.parentPath.isProgram()) {
-          this.movedStatements.push(path.node);
-        }
-      },
-      Program: {
-        enter(path) {
-          // If the body of the program consists of a single string literal,
-          // we want to intepret it as such and not as a a directive (that is,
-          // a "use strict"-like thing).
-          if (path.node.directives.length === 1 &&
-              path.node.directives[0].value.type === 'DirectiveLiteral') {
-            path.replaceWith(t.program([
-              t.expressionStatement(
-                t.stringLiteral(path.node.directives[0].value.value))
-            ]));
-          }
-        },
-        exit(path): void {
-          // Once we are done gathering all statements and variables,
-          // create a program that has a list of variables and the rest of the
-          // code inside an IIFE.
-          if (this.hasFinishedMoving) return;
-          this.hasFinishedMoving = true;
-          this.completionRecordId = path.scope.generateUidIdentifier('cr');
-          this.movedStatements.unshift(
-            t.variableDeclaration('var', [t.variableDeclarator(this.completionRecordId)]));
-          path.replaceWith(t.program([
-            ...this.variables.map(
-              v => t.variableDeclaration('var', [t.variableDeclarator(t.identifier(v))])),
-            ...this.functionDeclarations,
-            t.expressionStatement(t.callExpression(
-              t.arrowFunctionExpression(
-                [],
-                t.blockStatement(this.movedStatements)
-              ), []))]));
-        }
-      },
-      BlockStatement: {
-        exit(path): void {
-          if (!this.hasFinishedMoving) return;
-          // After creating the function, we look for completion records and
-          // turn them into return statements. This applies only to body of
-          // the single top-level function that we just created.
-          if (!path.parentPath.isArrowFunctionExpression()) return;
-          if (path.parentPath.getFunctionParent()) return;
-          if (this.addedCompletionRecords) return;
-          this.addedCompletionRecords = true;
-          const records = path.getCompletionRecords();
-          for (const record of records) {
-            // ExpressionWrapper = ExpressionStatement | ParenthesizedExpression
-            if (record.isExpressionWrapper() && !record.node.expression[excludeFromCompletion]) {
-              record.replaceWith(t.expressionStatement(
-                t.assignmentExpression('=', this.completionRecordId, record.node.expression)));
-            }
-          }
-          path.replaceWith(t.blockStatement([
-            ...path.node.body,
-            t.returnStatement(this.completionRecordId)
-          ]));
-        }
-      }
-    }
-  };
-};
-
-interface AsyncFunctionIdentifiers {
-  functionState: babel.types.Identifier;
-  synchronousReturnValue: babel.types.Identifier;
-  asynchronousReturnValue: babel.types.Identifier;
-  expressionHolder: babel.types.Identifier;
-  markSyntheticPromise: babel.types.Identifier;
-  isSyntheticPromise: babel.types.Identifier;
-  syntheticPromiseSymbol: babel.types.Identifier;
-  demangleError: babel.types.Identifier;
-  assertNotSyntheticPromise: babel.types.Identifier;
-}
-/**
- * The second step that performs the heavy lifting of turning regular functions
- * into maybe-async-maybe-not functions.
- *
- * This consists of two main components:
- *
- * 1. A function meta-wrapper, where an original (non-async) function is
- *    converted into an async function and is placed (as an "inner" function)
- *    into the body of another function and is wrapper. If the inner function
- *    returns (or throws) synchronously, the outer function also return
- *    synchronously. If not, the return value is marked as a Promise that should
- *    implicitly be awaited in other rewritten code.
- *
- * 2. An expression wrapper, which looks at expressions inside the original
- *    function body, and inserts code that dynamically decides whether to await
- *    the expressions it encounters or not at runtime.
- *
- * The README file has more complete code snippets.
- */
-export const makeMaybeAsyncFunctionPlugin = ({ types: t }: { types: typeof BabelTypes }): babel.PluginObj<{ file: babel.types.File }> => {
-  // We mark certain AST nodes as 'already visited' using these symbols.
-  function asNodeKey(v: any): keyof babel.types.Node { return v; }
-  const isGeneratedInnerFunction = asNodeKey(Symbol('isGeneratedInnerFunction'));
-  const isGeneratedHelper = asNodeKey(Symbol('isGeneratedHelper'));
-  const isOriginalBody = asNodeKey(Symbol('isOriginalBody'));
-  const isAlwaysSyncFunction = asNodeKey(Symbol('isAlwaysSyncFunction'));
-  // Using this key, we store data on Function nodes that contains the identifiers
-  // of helpers which are available inside the function.
-  const identifierGroupKey = '@@mongosh.identifierGroup';
-
-  // We fetch the symbol constructor as
-  //   Object.getOwnPropertySymbols(Array.prototype)[0].constructor
-  // because Symbol refers to BSONSymbol inside the target mongosh context.
-  // (This is the only mongosh-specific hack in here.)
-  const symbolConstructor = babel.template.expression(`
-    Object.getOwnPropertySymbols(Array.prototype)[0].constructor
-  `)();
-
-  const syntheticPromiseSymbolTemplate = babel.template.statement(`
-    const SP_IDENTIFIER = SYMBOL_CONSTRUCTOR.for("@@mongosh.syntheticPromise");
-  `);
-
-  const markSyntheticPromiseTemplate = babel.template.statement(`
-    function MSP_IDENTIFIER(p) {
-      return Object.defineProperty(p, SP_IDENTIFIER, {
-        value: true
-      });
-    }
-  `);
-
-  const isSyntheticPromiseTemplate = babel.template.statement(`
-    function ISP_IDENTIFIER(p) {
-      return p && p[SP_IDENTIFIER];
-    }
-  `);
-
-  const assertNotSyntheticPromiseTemplate = babel.template.statement(`
-    function ANSP_IDENTIFIER(p, s) {
-      if (p && p[SP_IDENTIFIER]) {
-        throw new Error('Result of expression "' + s + '" cannot be used in this context');
-      }
-      return p;
-    }
-  `);
-
-  const asyncTryCatchWrapperTemplate = babel.template.expression(`
-    async () => {
-      try {
-        ORIGINAL_CODE;
-      } catch (err) {
-        if (FUNCTION_STATE_IDENTIFIER === "sync") {
-          SYNC_RETURN_VALUE_IDENTIFIER = err;
-          FUNCTION_STATE_IDENTIFIER = "threw";
-        } else throw err;
-      } finally {
-        if (FUNCTION_STATE_IDENTIFIER !== "threw") FUNCTION_STATE_IDENTIFIER = "returned";
-      }
-    }
-  `);
-
-  const wrapperFunctionTemplate = babel.template.statements(`
-    let FUNCTION_STATE_IDENTIFIER = "sync",
-        SYNC_RETURN_VALUE_IDENTIFIER,
-        EXPRESSION_HOLDER_IDENTIFIER;
-
-    const ASYNC_RETURN_VALUE_IDENTIFIER = (ASYNC_TRY_CATCH_WRAPPER)();
-
-    if (FUNCTION_STATE_IDENTIFIER === "returned")
-      return SYNC_RETURN_VALUE_IDENTIFIER;
-    else if (FUNCTION_STATE_IDENTIFIER === "threw")
-      throw SYNC_RETURN_VALUE_IDENTIFIER;
-    FUNCTION_STATE_IDENTIFIER = "async";
-    return MSP_IDENTIFIER(ASYNC_RETURN_VALUE_IDENTIFIER);
-  `);
-
-  const awaitSyntheticPromiseTemplate = babel.template.expression(`(
-    ORIGINAL_SOURCE,
-    EXPRESSION_HOLDER = NODE,
-    ISP_IDENTIFIER(EXPRESSION_HOLDER) ? await EXPRESSION_HOLDER : EXPRESSION_HOLDER
-  )`, {
-    allowAwaitOutsideFunction: true
-  });
-
-  const assertNotSyntheticExpressionTemplate = babel.template.expression(`
-    ANSP_IDENTIFIER(NODE, ORIGINAL_SOURCE)
-  `);
-
-  const rethrowTemplate = babel.template.statement(`
-    try {
-      ORIGINAL_CODE;
-    } catch (err) {
-      throw err;
-    }
-  `);
-
-  // If we encounter an error object, we fix up the error message from something
-  // like `("a" , foo(...)(...)) is not a function` to `a is not a function`.
-  // For that, we look for a) the U+FEFF markers we use to tag the original source
-  // code with, and b) drop everything else in this parenthesis group (this uses
-  // the fact that currently, parentheses in error messages are nested at most
-  // two levels deep, which makes it something that we can tackle with regexps).
-  const demangleErrorTemplate = babel.template.statement(String.raw `
-    function DE_IDENTIFIER(err) {
-      if (Object.prototype.toString.call(err) === '[object Error]' &&
-          err.message.includes('\ufeff')) {
-        err.message = err.message.replace(/\(\s*"\ufeff(.+?)\ufeff"\s*,(?:[^\(]|\([^\)]*\))*\)/g, '$1');
-      }
-      return err;
-    }
-  `, { placeholderPattern: false, placeholderWhitelist: new Set(['DE_IDENTIFIER']) });
-
-  const returnValueWrapperTemplate = babel.template.expression(`(
-    SYNC_RETURN_VALUE_IDENTIFIER = NODE,
-    FUNCTION_STATE_IDENTIFIER === 'async' ? SYNC_RETURN_VALUE_IDENTIFIER : null
-  )`);
-
-  return {
-    pre(file: babel.types.File) {
-      this.file = file;
-    },
-    visitor: {
-      BlockStatement(path) {
-        // This might be a function body. If it's what we're looking for, wrap it.
-        if (!path.parentPath.isFunction()) return;
-        // Don't wrap things we've already wrapped.
-        if (path.parentPath.getData(identifierGroupKey)) return;
-        // Don't wrap the inner function we've created while wrapping another function.
-        if (path.parentPath.node[isGeneratedInnerFunction]) return;
-        // Don't wrap helper functions with async-rewriter-generated code.
-        if (path.parentPath.node[isGeneratedHelper]) return;
-
-        const originalSource = path.parent.start !== undefined ?
-          (this.file as any).code.slice(path.parent.start, path.parent.end) :
-          'function () { [unknown code] }';
-        // Encode using UTF-16 + hex encoding so we don't have to worry about
-        // special characters.
-        const encodedOriginalSource =
-          [...originalSource].map(char => char.charCodeAt(0).toString(16).padStart(4, '0')).join('');
-        const originalSourceNode = t.expressionStatement(
-          t.stringLiteral(`<async_rewriter>${encodedOriginalSource}</>`));
-
-        // A parent function might have a set of existing helper methods.
-        // If it does, we re-use the functionally equivalent ones.
-        const existingIdentifiers: AsyncFunctionIdentifiers | null =
-          path.findParent(path => !!path.getData(identifierGroupKey))?.getData(identifierGroupKey);
-
-        // Generate and store a set of identifiers for helpers.
-        const functionState = path.scope.generateUidIdentifier('fs');
-        const synchronousReturnValue = path.scope.generateUidIdentifier('srv');
-        const asynchronousReturnValue = path.scope.generateUidIdentifier('arv');
-        const expressionHolder = path.scope.generateUidIdentifier('ex');
-        const markSyntheticPromise = existingIdentifiers?.markSyntheticPromise ?? path.scope.generateUidIdentifier('msp');
-        const isSyntheticPromise = existingIdentifiers?.isSyntheticPromise ?? path.scope.generateUidIdentifier('isp');
-        const assertNotSyntheticPromise = existingIdentifiers?.assertNotSyntheticPromise ?? path.scope.generateUidIdentifier('ansp');
-        const syntheticPromiseSymbol = existingIdentifiers?.syntheticPromiseSymbol ?? path.scope.generateUidIdentifier('sp');
-        const demangleError = existingIdentifiers?.demangleError ?? path.scope.generateUidIdentifier('de');
-        const identifiersGroup: AsyncFunctionIdentifiers = {
-          functionState,
-          synchronousReturnValue,
-          asynchronousReturnValue,
-          expressionHolder,
-          markSyntheticPromise,
-          isSyntheticPromise,
-          assertNotSyntheticPromise,
-          syntheticPromiseSymbol,
-          demangleError
-        };
-        path.parentPath.setData(identifierGroupKey, identifiersGroup);
-
-        // We generate code that vaguely looks like and insert it at the top
-        // of the wrapper function:
-        // const syntheticPromise = Symbol.for('@mongosh.syntheticPromise');
-        // function markSyntheticPromise(promise) {
-        //   return Object.defineProperty(promise, syntheticPromise, { value: true });
-        // }
-        // function isSyntheticPromise(value) {
-        //   return value && value[syntheticPromise];
-        // }
-        // Note that the last check potentially triggers getters and Proxy methods
-        // and we may want to replace it by something a bit more sophisticated.
-        // All of the top-level AST nodes here are marked as generated helpers.
-        const commonHelpers = existingIdentifiers ? [] : [
-          Object.assign(
-            syntheticPromiseSymbolTemplate({
-              SP_IDENTIFIER: syntheticPromiseSymbol,
-              SYMBOL_CONSTRUCTOR: symbolConstructor
-            }),
-            { [isGeneratedHelper]: true }
-          ),
-        ];
-        const promiseHelpers = existingIdentifiers ? [] : [
-          ...commonHelpers,
-          Object.assign(
-            markSyntheticPromiseTemplate({
-              MSP_IDENTIFIER: markSyntheticPromise,
-              SP_IDENTIFIER: syntheticPromiseSymbol
-            }),
-            { [isGeneratedHelper]: true }
-          ),
-          Object.assign(
-            isSyntheticPromiseTemplate({
-              ISP_IDENTIFIER: isSyntheticPromise,
-              SP_IDENTIFIER: syntheticPromiseSymbol
-            }),
-            { [isGeneratedHelper]: true }
-          ),
-          Object.assign(
-            demangleErrorTemplate({
-              DE_IDENTIFIER: demangleError
-            }),
-            { [isGeneratedHelper]: true }
-          )
-        ];
-        const syncFnHelpers = [
-          ...commonHelpers,
-          Object.assign(
-            assertNotSyntheticPromiseTemplate({
-              ANSP_IDENTIFIER: assertNotSyntheticPromise,
-              SP_IDENTIFIER: syntheticPromiseSymbol
-            }),
-            { [isGeneratedHelper]: true }
-          )
-        ];
-
-        if (path.parentPath.node.async) {
-          // If we are in an async function, no async wrapping is necessary.
-          // We still want to have the runtime helpers available, and we add
-          // a re-throwing try/catch around the body so that we can perform
-          // error message adjustment through the CatchClause handler below.
-          path.replaceWith(t.blockStatement([
-            originalSourceNode,
-            ...promiseHelpers,
-            rethrowTemplate({
-              ORIGINAL_CODE: path.node.body
-            })
-          ]));
-          return;
-        }
-
-        // If we are in a non-async generator function, or a class constructor,
-        // we throw errors for implicitly asynchronous expressions, because there
-        // is just no good way to handle them (e.g.: What happens when you
-        // subclass a class with a constructor that returns asynchronously?).
-        if (path.parentPath.node.generator ||
-            (path.parentPath.isClassMethod() &&
-             path.parentPath.node.key.type === 'Identifier' &&
-             path.parentPath.node.key.name === 'constructor')) {
-          Object.assign(path.parentPath.node, { [isAlwaysSyncFunction]: true });
-          path.replaceWith(t.blockStatement([
-            originalSourceNode,
-            ...syncFnHelpers,
-            rethrowTemplate({
-              ORIGINAL_CODE: path.node.body
-            })
-          ]));
-          return;
-        }
-
-        const asyncTryCatchWrapper = Object.assign(
-          asyncTryCatchWrapperTemplate({
-            FUNCTION_STATE_IDENTIFIER: functionState,
-            SYNC_RETURN_VALUE_IDENTIFIER: synchronousReturnValue,
-            ORIGINAL_CODE: Object.assign(path.node, { [isOriginalBody]: true })
-          }),
-          { [isGeneratedInnerFunction]: true }
-        );
-
-        const wrapperFunction = wrapperFunctionTemplate({
-          FUNCTION_STATE_IDENTIFIER: functionState,
-          SYNC_RETURN_VALUE_IDENTIFIER: synchronousReturnValue,
-          ASYNC_RETURN_VALUE_IDENTIFIER: asynchronousReturnValue,
-          EXPRESSION_HOLDER_IDENTIFIER: expressionHolder,
-          MSP_IDENTIFIER: markSyntheticPromise,
-          ASYNC_TRY_CATCH_WRAPPER: asyncTryCatchWrapper
-        });
-
-        // Generate the wrapper function. See the README for a full code snippet.
-        path.replaceWith(t.blockStatement([
-          originalSourceNode,
-          ...promiseHelpers,
-          ...wrapperFunction
-        ]));
-      },
-      Expression: {
-        enter(path) {
-          // Minor adjustment: When we encounter a 'shorthand' arrow function,
-          // i.e. `(...args) => returnValueExpression`, we transform it into
-          // one with a block body containing a single return statement.
-          if (path.parentPath.isArrowFunctionExpression() && path.key === 'body') {
-            path.replaceWith(t.blockStatement([
-              t.returnStatement(path.node)
-            ]));
-          }
-        },
-        exit(path) {
-          // We have seen an expression. If we're not inside an async function,
-          // or a function that we explicitly marked as needing always-synchronous
-          // treatment, we don't care.
-          if (!path.getFunctionParent()) return;
-          if (!path.getFunctionParent().node.async &&
-              !path.getFunctionParent().node[isAlwaysSyncFunction]) return;
-          // identifierGroup holds the list of helper identifiers available
-          // inside this function.
-          let identifierGroup: AsyncFunctionIdentifiers;
-          if (path.getFunctionParent().node[isGeneratedInnerFunction]) {
-            // We are inside a generated inner function. If there is no node
-            // marked as [isOriginalBody] between it and the current node,
-            // we skip this (for example, this applies to the catch and finally
-            // blocks generated above).
-            if (!path.findParent(
-              path => path.isFunction() || !!path.node[isOriginalBody]
-            ).node[isOriginalBody]) {
-              return;
-            }
-
-            // We know that the outer function of the inner function has
-            // helpers available.
-            identifierGroup = path.getFunctionParent().getFunctionParent().getData(identifierGroupKey);
-            if (path.parentPath.isReturnStatement() && !path.node[isGeneratedHelper]) {
-              // If this is inside a return statement that we have not already handled,
-              // we replace the `return ...` with
-              // `return (_synchronousReturnValue = ..., _functionState === 'async' ? _synchronousReturnValue : null)`.
-              path.replaceWith(Object.assign(
-                returnValueWrapperTemplate({
-                  SYNC_RETURN_VALUE_IDENTIFIER: identifierGroup.synchronousReturnValue,
-                  FUNCTION_STATE_IDENTIFIER: identifierGroup.functionState,
-                  NODE: path.node
-                }),
-                { [isGeneratedHelper]: true }
-              ));
-              return;
-            }
-          } else {
-            // This is a regular async function. We also transformed these above.
-            identifierGroup = path.getFunctionParent().getData(identifierGroupKey);
-          }
-
-          // If there is a [isGeneratedHelper] between the function we're in
-          // and this node, that means we've already handled this node.
-          if (path.find(
-            path => path.isFunction() || !!path.node[isGeneratedHelper]
-          ).node[isGeneratedHelper]) {
-            return;
-          }
-
-          // Do not transform foo.bar() into (expr = foo.bar, ... ? await expr : expr)(),
-          // because that would mean that the `this` value inside the function is
-          // incorrect. This would be hard to get right, but it should be okay for now
-          // as we don't currently have to deal with situations in which functions
-          // themselves are something that we need to `await`.
-          // If we ever do, replacing all function calls with
-          // Function.prototype.call.call(fn, target, ...originalArgs) might be
-          // a feasible solution.
-          // Additionally, skip calls to 'eval', since literal calls to it
-          // have semantics that are different from calls to an expressio that
-          // evaluates to 'eval'.
-          if (path.parentPath.isCallExpression() &&
-              path.key === 'callee' &&
-              (path.isMemberExpression() || (path.isIdentifier() && path.node.name === 'eval'))) {
-            return;
-          }
-
-          // This, on the other hand, (e.g. the `foo.bar` in `foo.bar = 1`) is
-          // something that we wouldn't want to modify anyway, because it would
-          // break the assignment altogether.
-          if (path.parentPath.isAssignmentExpression() && path.key === 'left') return;
-          // ++ and -- count as assignments for our purposes.
-          if (path.parentPath.isUpdateExpression()) return;
-
-          // There are a few types of expressions that we can skip.
-          // We use this opt-out-list approach so that we don't miss any important
-          // expressions.
-          if (path.isLiteral() || // type is known to be non-Promise (here and below)
-              path.isArrayExpression() ||
-              path.isObjectExpression() ||
-              path.isFunctionExpression() ||
-              path.isArrowFunctionExpression() ||
-              path.isClassExpression() ||
-              path.isAssignmentExpression() || // sub-nodes are already awaited (ditto)
-              path.isBinaryExpression() ||
-              path.isConditionalExpression() ||
-              path.isLogicalExpression() ||
-              path.isSequenceExpression() ||
-              path.isParenthesizedExpression() ||
-              path.isUnaryExpression() ||
-              path.isSuper() || // Would probably break stuff
-              path.isAwaitExpression() || // No point in awaiting twice
-              path.parentPath.isAwaitExpression()) {
-            return;
-          }
-
-          const { expressionHolder, isSyntheticPromise, assertNotSyntheticPromise } = identifierGroup;
-          const prettyOriginalString = limitStringLength(
-            path.node.start !== undefined ?
-              (this.file as any).code.slice(path.node.start, path.node.end) :
-              '<unknown>', 24);
-
-          if (!path.getFunctionParent().node.async) {
-            // Transform expression `foo` into `assertNotSyntheticPromise(foo, 'foo')`.
-            path.replaceWith(Object.assign(
-              assertNotSyntheticExpressionTemplate({
-                ORIGINAL_SOURCE: t.stringLiteral(prettyOriginalString),
-                NODE: path.node,
-                ANSP_IDENTIFIER: assertNotSyntheticPromise
-              }),
-              { [isGeneratedHelper]: true }
-            ));
-            return;
-          }
-
-          // Transform expression `foo` into
-          // `('\uFEFFfoo\uFEFF', ex = foo, isSyntheticPromise(ex) ? await ex : ex)`
-          // The first part of the sequence expression is used to identify this
-          // expression for re-writing error messages, so that we can transform
-          // TypeError: ((intermediate value)(intermediate value) , (intermediate value)(intermediate value)(intermediate value)).findx is not a function
-          // back into
-          // TypeError: db.test.findx is not a function
-          // The U+FEFF markers are only used to rule out any practical chance of
-          // user code accidentally being recognized as the original source code.
-          // We limit the string length so that long expressions (e.g. those
-          // containing functions) are not included in full length.
-          const originalSource = t.stringLiteral(
-            '\ufeff' + prettyOriginalString + '\ufeff');
-          path.replaceWith(Object.assign(
-            awaitSyntheticPromiseTemplate({
-              ORIGINAL_SOURCE: originalSource,
-              EXPRESSION_HOLDER: expressionHolder,
-              ISP_IDENTIFIER: isSyntheticPromise,
-              NODE: path.node
-            }),
-            { [isGeneratedHelper]: true }
-          ));
-        }
-      },
-      CatchClause: {
-        exit(path) {
-          if (path.node[isGeneratedHelper] || !path.node.param || path.node.param.type !== 'Identifier') return;
-          const existingIdentifiers: AsyncFunctionIdentifiers | null =
-            path.findParent(path => !!path.getData(identifierGroupKey))?.getData(identifierGroupKey);
-          if (!existingIdentifiers) return;
-          // Turn `... catch (err) { ... }` into `... catch (err) { err = demangleError(err); ... }`
-          path.replaceWith(Object.assign(
-            t.catchClause(path.node.param,
-              t.blockStatement([
-                t.expressionStatement(
-                  t.assignmentExpression('=', path.node.param,
-                    t.callExpression(existingIdentifiers.demangleError, [path.node.param]))),
-                path.node.body
-              ])),
-            { [isGeneratedHelper]: true }
-          ));
-        }
-      }
-    }
-  };
-};
-
-function limitStringLength(input: string, maxLength: number) {
-  if (input.length <= maxLength) return input;
-  return input.slice(0, (maxLength - 5) * 0.7) +
-    ' ... ' +
-    input.slice(input.length - (maxLength - 5) * 0.3);
-}
 
 export default class AsyncWriter {
   step(code: string, plugins: babel.PluginItem[]): string {
